починил роутер и сделал исправления для мерджа
