+++ index.css
@@ -1,3 +1,9 @@
+@import url("/src/pages/main/main.css");
+@import url("/src/components/navbar/navbar.css");
+@import url("/src/components/header/header.css");
+@import url("/src/pages/main/components/card/card.css");
+
+
 * {
   margin: 0;
   padding: 0;
@@ -33,127 +39,18 @@ body {
 #root {
   justify-content: center;
   display: flex;
+  flex-direction: column;
+  width: 100%;
+  height: 100%;
 }
 
 .main {
   display: flex;
   margin: 86px 67px;
   gap: 12px;
-  height: 853px;
+  height: 100%;
   justify-content: center;
   align-items: center;
-  width: 1306px;
-}
-
-.navbar {
-  width: 378px;
-  background: #ffffffa8;
-  border-radius: 13px;
-  height: 100%;
-}
-
-.navbar__header {
-  background: #0b0c0c;
-  border-radius: 13px 13px 0px 0px;
-  padding: 20px 22px;
-  gap: 15px;
-}
-
-.navbar__header__person {
-  display: flex;
-  align-items: center;
-  gap: 15px;
-  justify-content: space-between;
-}
-
-.navbar__header__person__name {
-  /* font-family: Roboto Flex; */
-  font-size: 32px;
-  font-weight: 600;
-  line-height: 32px;
-  letter-spacing: 0em;
-  text-align: left;
-  color: #ffffff;
-}
-
-.navbar__header__person__image {
-  width: 70px;
-  height: 70px;
-  border-radius: 100px;
-}
-
-.navbar__menu {
-  padding: 20px 22px;
-  gap: 15px;
-}
-
-.navbar__menu__search {
-  width: 340px;
-  height: 34px;
-  padding: 0px 1px;
-  border-radius: 10px;
-  background: #dcdcdc;
-  display: flex;
-  align-items: center;
-}
-
-.navbar__menu__search__input {
-  background: #dcdcdc;
-  border: none;
-  border-radius: 10px;
-  padding: 2px 0 0 5px;
-  width: 100%;
-  height: 100%;
-  text-align: justify;
-}
-
-.navbar__menu__items {
-  gap: 10px;
-  display: flex;
-  flex-direction: column;
-  margin-top: 10px;
-}
-
-.navbar__item {
-  display: flex;
-  align-items: center;
-  gap: 15px;
-  padding: 15px 10px;
-  font-size: 24px;
-  font-weight: 500;
-  line-height: 24px;
-  letter-spacing: 0em;
-  text-align: left;
-  cursor: pointer;
-  background: white;
-  transition: background 0.5s;
-  border-radius: 10px;
-}
-
-.navbar__item__icon {
-  width: 45px;
-  height: 45px;
-  border-radius: 100px;
-}
-
-.navbar__item__info {
-  letter-spacing: 0em;
-  text-align: left;
-}
-
-.navbar__item__info__name {
-  font-family: Roboto Flex;
-  font-size: 16px;
-  font-weight: 600;
-  line-height: 16px;
-}
-
-.navbar__item__info__message {
-  font-family: Roboto Flex;
-  font-size: 14px;
-  font-weight: 300;
-  line-height: 14px;
-  color: #5d5d5d;
 }
 
 .content {
@@ -176,129 +73,4 @@ body {
   transform-style: preserve-3d;
   /* position: relative; */
   /* margin-top: 13vh; */
-}
-
-.card {
-  width: 100%;
-  height: 100%;
-  position: absolute;
-  border-radius: 13px;
-  overflow: hidden;
-  transform: translateZ(calc(-30px * var(--i)));
-    /* translateY(calc(-20px * var(--i)));  */
-    /* rotate(calc(-4deg * var(--i))); */
-  filter: drop-shadow(2px 2px 20px rgba(0, 0, 0, 0.5));
-  cursor: pointer;
-  user-select: none;
-  transition: transform 0.5s;
-  background: white;
-}
-
-.card__image {
-  width: 364px;
-  height: 571px;
-  border-radius: 13px;
-  object-fit: cover;
-  object-position: 50% 50%;
-}
-.card__content {
-  /* position: absolute;
-  bottom: 0;
-  left: 0; */
-  /* width: 100%;
-  height: 15%; */
-  width: 237px;
-  height: 86px;
-  padding: 20px, 0px, 20px, 20px;
-  gap: 7px;
-
-  /* padding: 20px; */
-  /* background: linear-gradient(0deg, rgba(0, 0, 0, 0.5), rgba(0, 0, 0, 0)); */
-  color: black;
-  font-size: 2.5vh;
-  font-weight: 700;
-  /* text-align: center; */
-  /* transform: translateY(100%); */
-  transition: transform 0.5s;
-  padding: 0 6px;
-}
-
-.card__name {
-  /* padding: 10px 0; */
-  /* font-size: 3vh;
-  font-weight: 600; */
-  font-family: Roboto Flex;
-  font-size: 24px;
-  font-weight: 500;
-  line-height: 24px;
-  letter-spacing: 0em;
-  text-align: left;
-  transition: transform 0.5s;
-}
-
-.card__description {
-  /* padding: 3px;
-  font-size: 2vh;
-  font-weight: 400;
-  color: #8d8d8d;
-  transition: transform 0.5s; */
-  font-family: Roboto Flex;
-  font-size: 15px;
-  font-weight: 500;
-  line-height: 15px;
-  letter-spacing: 0em;
-  text-align: left;
-  transition: transform 0.5s;
-}
-
-#like,
-#dislike {
-  font-size: 6vh;
-  border-radius: 50%;
-  padding: 20px;
-  position: relative;
-  z-index: 1;
-  animation-name: animation1;
-  animation-duration: 1s;
-  animation-timing-function: ease-in-out;
-  animation-fill-mode: both;
-  /* don't run animation at the beginning */
-  animation-play-state: paused;
-}
-
-#like.trigger,
-#dislike.trigger {
-  animation-name: animation2;
-}
-
-#like {
-  color: red;
-  background-color: rgba(255, 255, 255, 0.5);
-}
-
-#dislike {
-  color: #ccc;
-  background-color: rgba(0, 0, 0, 0.5);
-}
-
-@keyframes animation1 {
-  0%,
-  100% {
-    opacity: 0.2;
-  }
-
-  50% {
-    opacity: 1;
-  }
-}
-
-@keyframes animation2 {
-  0%,
-  100% {
-    opacity: 0.2;
-  }
-
-  50% {
-    opacity: 1;
-  }
 }
