+++ index.html
@@ -7,16 +7,15 @@
     <!-- <link rel="stylesheet" href="index.css" /> -->
   </head>
   <body>
-    <div class="header">
+    <!-- <div class="header">
       <div class="header__logo">
-        <!-- <img src="img/logo.png" alt="SportBro"> -->
       </div>
       <div class="header__menu">
         <a class="header__menu-link" data-link href="/">Главная</a>
         <a class="header__menu-link" data-link href="/register">Регистрация</a>
         <a class="header__menu-link" data-link href="/login">Вход</a>
       </div>
-    </div>  
+    </div>   -->
     <div id="root"></div>
   </body>
 </html>
