+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "test": "test",
-    "start": "npm run build && node server/server.js",
+    "start": "npm run build && node server/server.mjs",
     "build": "webpack --mode production"
   },
   "repository": {
