+++ src/components/header/header.hbs
@@ -1,6 +1,5 @@
   <div class="header">
     <div class="header__logo">
-      <img src="img/logo.png" alt="SportBro">
     </div>
     <div class="header__menu">
       <a class="header__menu-link" data-link href="/">Главная</a>
