+++ src/pages/main/main.hbs
@@ -1,3 +1,4 @@
+{{> components/header/header}}
 <div class="main">
     {{!-- {{> partial/navbar/navbar}} --}}
     {{> components/navbar/navbar}}
