--- src/partial/navbar/navbar.hbs
@@ -1,8 +1,0 @@
-  <div class="navbar">
-    <div class="navbar__item">1</div>
-    <div class="navbar__item">2</div>
-    <div class="navbar__item">3</div>
-    <div class="navbar__item">4</div>
-    <div class="navbar__item">5</div>
-    <div class="navbar__item">6</div>
-  </div>
