+++ webpack.config.js
@@ -3,6 +3,10 @@ const path = require('path');
 
 module.exports = {
     entry: './index.js',
+        output: {
+        path: path.resolve(__dirname, 'dist'),
+        filename: '[name].js',
+    },
     module: {
         rules: [
             {test: /\.svg$/, use: 'svg-inline-loader'},
@@ -16,10 +20,6 @@ module.exports = {
         },
             {test: /\.html$/, use: ['html-loader']},
         ],
-    },
-    output: {
-        path: path.resolve(__dirname, 'dist'),
-        filename: '[name].js',
     },
     plugins: [
         new HtmlWebpackPlugin(
